Add to string method to inventory

+++ src/users/StudentController.java
@@ -5,6 +5,7 @@ import shop.*;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Set;
 
 public class StudentController extends UserController{
     StudentModel student;
@@ -14,7 +15,7 @@ public class StudentController extends UserController{
     public StudentController(StudentModel studentModel){
         student = studentModel;
         view = new StudentView();
-        updateInventory();
+        updateStudentInventory();
     }
 
     private void showMyWallet(){
@@ -37,14 +38,14 @@ public class StudentController extends UserController{
 
     private void showMyInventory() { view.displayInventory(student.getInventory()); }
 
-    private void showTeamInventory() { view.displayInventory(student.getTeam().getInventory()); }
+//    private void showTeamInventory() { view.displayInventory(student.getTeam().getInventory()); }
 
 //    public void removeFromInventory(ArtifactModel artifact) {student.getInventory().remove(artifact); }
 
     private void updateStudentInventory() {
+        student.getInventory().setStock();
         shopDAO = new ShopDAO();
-        shopDAO.loadStudentInventory(student.getId());
-        shopDAO.saveInventory(student.getId(), student.getInventory());
+        shopDAO.saveInventory(student.getId(), "students_artifacts", student.getInventory());
     }
 
 //    private void updateTeamInventory() {
@@ -62,20 +63,24 @@ public class StudentController extends UserController{
 
     private void useArtifacts() {
         showMyInventory();
-        if(student.getInventory().isEmpty()){
+        if(student.getInventory().getStock().isEmpty()){
             view.displayMessage("Sorry, You have nothing to use!");
             view.handlePause();
         } else {
             int id = view.getNumber("Enter artifact id: ");
-            for(ArtifactModel artifact : student.getInventory()) {
-                if(id == artifact.getId()) {
-
-                    student.getInventory().remove(artifact);
+            Set<ArtifactModel> artifacts = student.getInventory().getStock().keySet();
+            for(ArtifactModel artifact : artifacts) {
+                if(id == artifact.getId() && (Integer) student.getInventory().getStock().get(artifact) == 1) {
+                    student.getInventory().removeArtifact(artifact);
                     view.displayMessage("Artifact used!");
                     shopDAO.deleteFromInventory(student.getId(), id);
                     shopDAO.saveStudentTransaction(student.getId(), artifact.getId());
-                    updateInventory();
+                    updateStudentInventory();
                     break;
+                }
+                else {
+                    student.getInventory().decreaseQuantity(artifact);
+                    updateStudentInventory();
                 }
             }
         }
@@ -103,7 +108,7 @@ public class StudentController extends UserController{
 
                 case "1":
                     executeShopping();
-                    updateInventory();
+                    updateStudentInventory();
                     break;
                 case "2":
                     showMyInventory();
@@ -114,9 +119,9 @@ public class StudentController extends UserController{
                 case "4":
                     showStudentsFromMyTeam();
                     break;
-                case "5":
-                    showTeamInventory();
-                    break;
+//                case "5":
+//                    showTeamInventory();
+//                    break;
                 case "0":
                     isDone = true;
                     break;
